agregar cors al middelware

+++ src/controllers/productController.ts
@@ -1,33 +1,41 @@
 import { Request, Response } from "express"
 import Product from "../models/Product";
 
-interface Props {
-    req: Request;
-    res: Response;
-
-}
 
 
-export const getAllProducts = async ({ req, res }: Props): Promise<void> => {
+export const getAllProducts = async (req: Request, res: Response): Promise<void> => {
 
     try {
         const product = await Product.find();
         res.status(200).json(product);
+        console.log('object')
     } catch (error) {
         console.log(error)
-
     }
 };
 
-export const createProduct = async ({ req, res }: Props): Promise<void> => {
-    const { id, nombre, img, descripcion, precio, categoria, cantidad } = req.body;
+export const createProduct = async (req: Request, res: Response): Promise<void> => {
+   
     try {
-        const product = new Product({ id, nombre, img, descripcion, precio, categoria, cantidad });
+        if (!req.user?.isAdmin) {
+            res.status(401).json({
+                msg: 'No autorizado'
+            });
+            return
+        }
+    } catch (error) {
+        throw new Error('Failed token');
+    }
+    const { id, name, img, describe, price, category, stock } = req.body;
 
-        const productDB = await Product.findOne({ nombre });
+
+    try {
+        const product = new Product({ id, name, img, describe, price, category, stock });
+
+        const productDB = await Product.findOne({ name });
         if (productDB) {
             res.status(400).json({
-                msg: `El producto ${nombre} ya existe`
+                msg: `El producto ${name} ya existe`
             });
             return;
         }
@@ -40,7 +48,7 @@ export const createProduct = async ({ req, res }: Props): Promise<void> => {
     }
 };
 
-export const getOneProduct = async ({ req, res }: Props): Promise<void> => {
+export const getOneProduct = async (req: Request, res: Response): Promise<void> => {
     const { id } = req.params;
     try {
         const product = await Product.findById(id);
@@ -59,7 +67,13 @@ export const getOneProduct = async ({ req, res }: Props): Promise<void> => {
     }
 };
 
-export const updateProduct = async ({ req, res }: Props): Promise<void> => {
+export const updateProduct = async (req: Request, res: Response): Promise<void> => {
+    if (!req.user?.isAdmin) {
+        res.status(401).json({
+            msg: 'No autorizado'
+        });
+        return
+    }
     const { id } = req.params;
     const { _id, ...resto } = req.body;
     try {
@@ -71,12 +85,18 @@ export const updateProduct = async ({ req, res }: Props): Promise<void> => {
     }
 };
 
-export const deleteProduct = async ({ req, res }: Props): Promise<void> => {
+export const deleteProduct = async (req: Request, res: Response): Promise<void> => {
+    if (!req.user?.isAdmin) {
+        res.status(401).json({
+            msg: 'No autorizado'
+        });
+        return
+    }
     const { id } = req.params;
     try {
         const product = await Product.findByIdAndDelete(id);
         if (!product) {
-             res.status(404).json({
+            res.status(404).json({
                 msg: `Producto con el id ${id} no encontrado`
             });
             return
@@ -84,7 +104,7 @@ export const deleteProduct = async ({ req, res }: Props): Promise<void> => {
         res.status(200).json({
             msg: `Producto con el id ${id} eliminado`
         });
-    } catch (error:any) {
+    } catch (error: any) {
         console.error(error);
         res.status(500).json({
             msg: 'Error al eliminar el producto',
